Ok now we have something that works
