+++ build.gradle
@@ -10,6 +10,6 @@ allprojects {
   project.ext {
     media3Version = '1.0.2'
 //    coreVersion = "0.7.4"
-    coreVersion = 'dev-rendition-list-490764e'
+    coreVersion = 'dev-rendition-list-5b678be'
   }
 }
